restyled, refactored, reuse components, fix naming
