+++ src/components/StockDetails/StockDataGraph.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import { Box, useColorModeValue } from '@chakra-ui/react';
+import { useColorModeValue } from '@chakra-ui/react';
 import { defaults, Line } from 'react-chartjs-2';
 import { useStockSymbol } from '../../contexts/useStockInfo';
 import LoadingSpinner from '../LoadingSpinner';
+import CustomBox from '../CustomBox';
 
 const StockDataGraph = () => {
     const { graphData, isStockGraphLoading } = useStockSymbol();
@@ -10,18 +11,13 @@ const StockDataGraph = () => {
     defaults.borderColor = useColorModeValue('#CCC', '#555');
 
     return (
-        <Box
-            mt={10}
-            p={4}
-            bgColor={useColorModeValue('brand.100', 'brand.700')}
-            borderRadius="lg"
-        >
+        <CustomBox bgColor={useColorModeValue('brand.100', 'brand.700')}>
             {isStockGraphLoading ? (
                 <LoadingSpinner />
             ) : (
                 <Line data={graphData} />
             )}
-        </Box>
+        </CustomBox>
     );
 };
