+++ src/pages/StockInfo.jsx
@@ -4,7 +4,7 @@ import { useColorModeValue } from '@chakra-ui/react';
 import StockInformation from '../components/StockDetails/StockInformation';
 import StockNews from '../components/StockDetails/StockNews';
 import StockPredictionDetails from '../components/StockDetails/StockPredictionDetails';
-import StockDataGraph from '../components/StockDetails/StockPredictionOverview';
+import StockDataGraph from '../components/StockDetails/StockDataGraph';
 import { usePrediction } from '../contexts/usePredictions';
 import { useStockSymbol } from '../contexts/useStockInfo';
 import { useStockNews } from '../contexts/useStockNews';
